// frontend/src/app/model/report.ts
export class Report{
    selectedDate: String;
    selectedTime: String;
    doctorUsername: String;
    doctorFirstname: String;
    doctorLastname: String;
    doctorSpecialization: String;
    reason: String;
    diagnosis: String;
    therapy: String;
    nextDate: String;
    patientUsername: String;
    patientFirstname: String;
    patientLastname: String;
}